Code refactor
- Remove EventEmitter inheritance
- Split Queue and Parser to two files
- More comments
- Remove timeout argument from RedisQueue
