+++ task-queue/index.js
@@ -1,5 +1,7 @@
 const Queue = require('./lib/queue');
+const Parser = require('./lib/parser');
 
 module.exports = {
-  ...Queue
+  ...Queue,
+  ...Parser
 };
