+++ task-queue/lib/queue.test.js
@@ -1,14 +1,15 @@
 const Redis = require('ioredis');
 const {
-  RedisQueue,
-  JsonParser,
-  ParserError
+  RedisQueue
 } = require('./queue');
 
+// Url to redis databse with user,password,host and port
 const REDIS_URL = 'redis://127.0.0.1:6379';
-const QUEUE_NAME = 'test';
+// Name of testing queue
+const QUEUE_NAME = 'queue:test';
+// Timeout for queue
 const QUEUE_TIMEOUT = 30;
-const QUEUE_PREFIX = 'queue:';
+// Redis config for queue
 const QUEUE_REDIS_CONFIG = {
   host: '127.0.0.1',
   port: 6379,
@@ -16,7 +17,9 @@ const QUEUE_REDIS_CONFIG = {
 };
 
 describe('RedisQueue', () => {
+  // Database connection
   let db;
+  // Test message
   const message = {
     type: 'test',
     args: [
@@ -24,27 +27,32 @@ describe('RedisQueue', () => {
       'io'
     ]
   };
+  // Queue
   let queue;
 
+  // Create redis connection and queue
   beforeAll(() => {
     db = new Redis(REDIS_URL);
     queue = new RedisQueue({
       queueName: QUEUE_NAME,
-      prefix: QUEUE_PREFIX,
       timeput: QUEUE_TIMEOUT,
       redisConfig: QUEUE_REDIS_CONFIG
     });
   });
 
+  // Close all connections
   afterAll(() => {
     db.quit();
     queue.quit();
   });
 
+  // Send message via queue and check it via another connection
   it('should send a message to queue', async () => {
     try {
+      // Send message via queue
       await queue.sendMessage(message);
-      const response = await db.lindex(`${QUEUE_PREFIX}${QUEUE_NAME}`, 0);
+      // Receive it via plain connection
+      const response = await db.lindex(QUEUE_NAME, 0);
 
       expect(JSON.parse(response)).toEqual(message);
     } catch (e) {
@@ -53,6 +61,7 @@ describe('RedisQueue', () => {
     }
   });
 
+  // Pop created message from Redis queue
   it('should receive a message from the queue', async () => {
     try {
       const received = await queue.receiveMessage();
@@ -62,28 +71,5 @@ describe('RedisQueue', () => {
       console.error(e);
       return e;
     }
-  });
-});
-
-describe('JsonParser', () => {
-  const obj = {
-    type: 'test',
-    args: [ 'test' ],
-    flag: true,
-    num: 32197361
-  };
-  const corruptStr = '{"type": "test" ,}';
-  let parser = JsonParser;
-
-  it('should stringify an object', () => {
-    expect(parser.prepare(obj)).toBe(JSON.stringify(obj));
-  });
-
-  it('should parse an object', () => {
-    expect(parser.parse(JSON.stringify(obj))).toEqual(obj);
-  });
-
-  it('should throw ParserError on corrupt string', () => {
-    expect(() => parser.parse(corruptStr)).toThrowError(ParserError);
   });
 });
